// packages/udecode/depset/src/utils/logger.d.ts
export declare const highlighter: {
    error: typeof import("kleur/colors").print;
    info: typeof import("kleur/colors").print;
    success: typeof import("kleur/colors").print;
    warn: typeof import("kleur/colors").print;
};
export declare const logger: {
    break(): void;
    error(...args: unknown[]): void;
    info(...args: unknown[]): void;
    log(...args: unknown[]): void;
    success(...args: unknown[]): void;
    warn(...args: unknown[]): void;
};
//# sourceMappingURL=logger.d.ts.map